+++ apps/ced-portal-fe/src/pages/OpportunityDetail/components/OpportunityDetailCard.tsx
@@ -10,6 +10,12 @@ import {
 } from '@mui/material';
 import type { AdminOpportunityDetail } from '../../../features/opportunities/types';
 import { DetailSection } from './DetailSection';
+import {
+  getBenefitsDetailData,
+  getLocalizedMetadataDetailsMultipleKeys,
+} from '../../../utils';
+import { useAppSelector } from '../../../hooks';
+import { selectActiveFormLanguage } from '../../../features/opportunityCreation/selectors';
 
 interface OpportunityDetailCardProps {
   detail: AdminOpportunityDetail;
@@ -18,10 +24,29 @@ interface OpportunityDetailCardProps {
 export const OpportunityDetailCard = ({
   detail,
 }: Readonly<OpportunityDetailCardProps>) => {
+  const [description, condition] = getLocalizedMetadataDetailsMultipleKeys(
+    detail.localizedMetadata,
+    ['description', 'condition'],
+  );
+
+  const activeLanguage = useAppSelector(selectActiveFormLanguage);
+
+  const beneficiaryBenefitFields = getBenefitsDetailData(
+    detail.beneficiaryBenefit,
+    activeLanguage,
+  );
+
+  const caregiverBenefitFields = getBenefitsDetailData(
+    detail.caregiverBenefit,
+    activeLanguage,
+  );
+
   const mainFields = [
-    ...(detail.operatorName
-      ? [{ label: 'Ente', value: detail.operatorName }]
-      : []),
+    ...beneficiaryBenefitFields,
+    {
+      label: 'Descrizione',
+      value: description,
+    },
     { label: 'Categoria', value: detail.categoryTitle },
     {
       label: 'Inizio validità',
@@ -35,6 +60,7 @@ export const OpportunityDetailCard = ({
           },
         ]
       : []),
+    ...(condition ? [{ label: 'Condizioni', value: condition }] : []),
     ...(detail.url ? [{ label: 'URL', value: detail.url }] : []),
   ];
 
@@ -72,14 +98,7 @@ export const OpportunityDetailCard = ({
               </Typography>
             </Box>
             <Divider />
-            <DetailSection
-              fields={[
-                {
-                  label: 'Beneficio accompagnatore',
-                  value: detail.caregiverBenefit?.type ?? '',
-                },
-              ]}
-            />
+            <DetailSection fields={caregiverBenefitFields} />
           </>
         )}
       </AccordionDetails>