+++ packages/io-core-adapter-ar/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @pagopa/io-core-adapter-ar
 
+## 0.0.6
+
+### Patch Changes
+
+- 2870ce6: add onboarding manager info
+
+## 0.0.5
+
+### Patch Changes
+
+- 7b0924a: Support multi-status filtering for onboarding list API.Expand onboarding detail payload with full AR-backed fields.
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [40aaab7]
+  - @pagopa/io-core-domain@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes