+++ packages/io-core-adapter-fims/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pagopa/io-core-adapter-fims
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [40aaab7]
+  - @pagopa/io-core-domain@0.0.4
+  - @pagopa/io-core-adapter-fastify@0.0.5
+
 ## 0.0.3
 
 ### Patch Changes